# openapi.yaml
openapi: 3.0.3
info:
  title: Sorokchat
  version: 1.0.0
  description: Український захищений месенджер
components:
  securitySchemes:
    bearerAuthentication:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    GetUserPayload:
      type: object
      required:
        - id
        - login
        - displayName
      properties:
        id:
          type: number
          description: Унікальний ідентифікатор
          example: 1
        login:
          type: string
          description: Логін
          example: sorokovsky_andrey
        displayDisplay:
          type: string
          description: Видиме ім'я
          example: Сороковський Андрій
...
    ErrorPayload:
      type: object
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
          description: Статус код
          example: 400
        message:
          type: string
          description: Повідомлення
          additionalProperties: true
...
  responses:
...
    UnauthorizedResponse:
      description: Користувач не авторизований
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    ProfileResponse:
      description: Успішне отримання
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetUserPayload'
servers:
  - url: http://localhost:8080
paths:
...
  /authorization/profile: 
    get:
      security:
        - bearerAuthentication: []
      summary: Профіль
      description: Отримання даних авторизованого користувача
      operationId: profile
      tags:
        - Авторизація
      responses:
        200:
          $ref: "#/components/responses/ProfileResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
...